Implement cube rendering
Draw a shaded cube in the style of kmscube.

The cube is static and not transformed identically to the kmscube example yet. This commit establishes basing rendering and commits a shader that tries to follow the same patterns used in the kmscube example. TODOs note where future improvements need to be made.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>