Proper aspect ratio handling
Extract the Uniforms generation code to a separate function. Currently it takes the step and aspect ratio as arguments. Eventually it will become an &self function which uses the current step number to generate the correction transformation.

When a resize event is received from Winit, call the new .resize() function on the Cube

Add a .resize() function to the cube which recalculates the uniforms and writes the updated buffer.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>