Add clap to xtask
Add clap for argument parsing in the xtask script.

Use the builder pattern instead of the derive pattern for slightly better compile times.

clap is complete overkill for xtask, but I would like to make xtask user-friendly with clap's help messages, colors, and other amenities. I also anticipate using clap in the main wgpucube bin so using it here for consistency is a good idea.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>