Add tracing
Add the tracing library for logging and future performance measurement.

Tracing is overkill for such a simple library but it's flexible and familiar.

The standard tracing subscriber library won't work in wasm environments, so wasm_tracing is used as a bridge. The wasm_tracing crate appears to be minimal and easy to review, though it's forked from another crate that has been unmaintained for years. Ideally this can all be cleaned up in the future when proper wasm support is added to tracing.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>