Initial wasm init and rendering
First pass at supporting wasm.

Web platforms are not allowed to block but the process of initializing wgpu is async. As a workaround, wasm_bindgen_futures is used to spawn a future which sends a Winit UserEvent to the application when initialization is complete on wasm platforms. Other platforms retain the direct initialization in .resumed().

I have added TODO notes where further improvement is needed. Notably, some Winit events are dropped while waiting for initialization. These could be queued up and handled after initialization.

This code is good enough to commit as a wasm checkpoint before moving on to finer details.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>