Workaround iOS request_redraw bug
Winit v0.30 (and likely earlier) has a bug where calling window.request_redraw() during a WindowEvent::RedrawRequested does not properly queue the redraw request.

As a result, the standard method of requesting a redraw after handling each RedrawRequested event does not work. Only a single frame is drawn.

This adds a workaround and documents it with comments and a section in the README.md

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>