xtask run-ios command
Add a `run-ios` subcommand to xtask that will build the project for the iOS simulator, launch an iOS simulator, and run the compiled program.

This workflow could use a lot of additional polish to check preconditions and handle edge cases, such as requiring macOS and confirming that the necessary binaries are installed. I added some TODO notes for the future improvements. For now, this is a quick way to get the app running on iOS.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>