Android support
Add an Android specific entry point and instructions for running with cargo-apk

Android apps must be built as libraries, so a separate entry point has been created to build as a lib. This is currently in the shared main.rs file which produces a Cargo warning. A future refactor can restructure the app so that shared components are in a separate file and main.rs and lib.rs act as entry points.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>