Add Android to CI
Android builds as a library, not a binary. The CI must be updated to add a `clippy_targets` option for each platform.

On non-Android platforms the `--bins` option is used to check the wgpucube bin and xtask.

On Android, the clippy targets are `--lib --bin xtask` to check the lib and only the xtask binary.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>