ci: Don't install default toolchain with rustup
Don't install the default toolchain when installing rustup on ubuntu-slim builders.

rustup was installed by default on the larger GitHub workers, but on ubuntu-slim the rustup toolchain install command is invoked. The previous version would install a default toolchain as this is the default behavior of rustup.

This changes the rustup install command to not install a default toolchain. This is important because the workflows install the correct toolchain in the next step. Should save some CI time.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>