Push wasm artifacts to GH Pages repo
Rework the wasm build and deploy workflow to push the updated wasm artifacts to the GitHub Pages repo.

This is ultimately cleaner than the previous method of uploading the artifacts and then triggering a workflow on the other repo. Only a single personal access token is necessary for this workflow, as opposed to the two tokens required with the previous system.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>