Use expect instead of allow for lints
Use the new expect directive from Rust 1.81 to be more explicity about where clippy warnings should be ignored. Using an expect directive tied to a specific platform makes it more explicit that the lint is expected and also gives a warning if the need for the lint exception is later removed.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>