Add error overlay in web view
Introduce an error overlay into the web version. The overlay is hidden with CSS by default.

A new error function in util.rs can unhide the overlay and show error text.

Add anyhow and allow the Context::new() function to return an anyhow::Result with any errors. These can be displayed on the overlay.

This allows webgpu related startup failures to be displayed in the web page itself, in addition to the log lines. Showing errors in the web page is much more helpful to casual visitors.

Signed-off-by: Tom Stokes <tomstokes@radixengineering.com>